Restructure some command attributes
