+++ Pepper/Commands/Debugging/DebuggingCommands.cs
@@ -1,6 +1,6 @@
 using Disqord.Bot.Commands;
 using Pepper.Structures;
-using Pepper.Structures.Commands;
+using Pepper.Structures.CommandAttributes.Metadata;
 
 namespace Pepper.Commands.Debugging
 {
