+++ Pepper/Commands/Osu/Rank.cs
@@ -10,7 +10,8 @@ using osu.Game.Rulesets;
 using Pepper.Commons.Osu;
 using Pepper.Database.OsuUsernameProviders;
 using Pepper.Services.Osu;
-using Pepper.Structures.Commands;
+using Pepper.Structures;
+using Pepper.Structures.CommandAttributes.Checks;
 using Pepper.Structures.External.Osu.Extensions;
 using Qmmands;
 using Qmmands.Text;
