+++ Pepper/Commands/Osu/Score.cs
@@ -9,7 +9,8 @@ using osu.Game.Scoring;
 using Pepper.Commons.Osu;
 using Pepper.Database.OsuUsernameProviders;
 using Pepper.Services.Osu;
-using Pepper.Structures.Commands;
+using Pepper.Structures;
+using Pepper.Structures.CommandAttributes.Metadata;
 using Pepper.Structures.External.Osu;
 using Pepper.Utilities.Osu;
 using Qmmands;
