+++ Pepper/Commands/Reserved/WhitelistCommand.cs
@@ -5,7 +5,8 @@ using Microsoft.Extensions.DependencyInjection;
 using Pepper.Database;
 using Pepper.Services;
 using Pepper.Structures;
-using Pepper.Structures.Commands;
+using Pepper.Structures.CommandAttributes.Checks;
+using Pepper.Structures.CommandAttributes.Metadata;
 using Qmmands;
 using Qmmands.Text;
 using Command = Pepper.Structures.Command;
