+++ Pepper/Structures/Bot/Bot.cs
@@ -6,7 +6,7 @@ using Disqord.Bot;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using Pepper.Structures.Commands;
+using Pepper.Structures.CommandAttributes.Metadata;
 using Qmmands.Text;
 using Qmmands.Text.Default;
