--- Pepper/Structures/Commands/HiddenAttribute.cs
@@ -1,7 +1,0 @@
-namespace Pepper.Structures.Commands
-{
-    /// <summary>
-    /// Specify this on modules/commands to hide them from everyone except owners in <see cref="Pepper.Commands.General.Help"/> results.
-    /// </summary>
-    public class HiddenAttribute : System.Attribute { }
-}
